✨ Promote Couchbase to native mode Fixes #1188
