+++ docs/modules/ROOT/pages/reference/extensions/couchbase.adoc
@@ -4,16 +4,16 @@
 :page-aliases: extensions/couchbase.adoc
 :linkattrs:
 :cq-artifact-id: camel-quarkus-couchbase
-:cq-native-supported: false
-:cq-status: Preview
-:cq-status-deprecation: Preview
+:cq-native-supported: true
+:cq-status: Stable
+:cq-status-deprecation: Stable
 :cq-description: Query Couchbase Views with a poll strategy and/or perform various operations against Couchbase databases.
 :cq-deprecated: false
 :cq-jvm-since: 1.0.0
-:cq-native-since: n/a
+:cq-native-since: 2.6.0
 
 [.badges]
-[.badge-key]##JVM since##[.badge-supported]##1.0.0## [.badge-key]##Native##[.badge-unsupported]##unsupported##
+[.badge-key]##JVM since##[.badge-supported]##1.0.0## [.badge-key]##Native since##[.badge-supported]##2.6.0##
 
 Query Couchbase Views with a poll strategy and/or perform various operations against Couchbase databases.
 
@@ -24,6 +24,10 @@ Query Couchbase Views with a poll strategy and/or perform various operations aga
 Please refer to the above link for usage and configuration details.
 
 == Maven coordinates
+
+https://code.quarkus.io/?extension-search=camel-quarkus-couchbase[Create a new project with this extension on code.quarkus.io, window="_blank"]
+
+Or add the coordinates to your existing project:
 
 [source,xml]
 ----
