+++ docs/modules/ROOT/partials/reference/components/couchbase.adoc
@@ -2,11 +2,11 @@
 // This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 :cq-artifact-id: camel-quarkus-couchbase
 :cq-artifact-id-base: couchbase
-:cq-native-supported: false
-:cq-status: Preview
+:cq-native-supported: true
+:cq-status: Stable
 :cq-deprecated: false
 :cq-jvm-since: 1.0.0
-:cq-native-since: n/a
+:cq-native-since: 2.6.0
 :cq-camel-part-name: couchbase
 :cq-camel-part-title: Couchbase
 :cq-camel-part-description: Query Couchbase Views with a poll strategy and/or perform various operations against Couchbase databases.
