+++ integration-tests-jvm/pom.xml
@@ -52,7 +52,6 @@
         <module>coap</module>
         <module>cometd</module>
         <module>corda</module>
-        <module>couchbase</module>
         <module>djl</module>
         <module>dns</module>
         <module>drill</module>
