+++ integration-tests/couchbase/src/main/java/org/apache/camel/quarkus/component/couchbase/it/CouchbaseResource.java
